Python-Producer/requirements.txt
pip==21.1.2
setuptools==57.0.0
kafka-python==2.0.2